nftMarketplaceSubgraph/tests/erc-721-factory.test.ts
import {
  assert,
  describe,
  test,
  clearStore,
  beforeAll,
  afterAll
} from "matchstick-as/assembly/index"
import { BigInt, Address } from "@graphprotocol/graph-ts"
import { NftContractCreated } from "../generated/schema"
import { NftContractCreated as NftContractCreatedEvent } from "../generated/ERC721Factory/ERC721Factory"
import { handleNftContractCreated } from "../src/erc-721-factory"
import { createNftContractCreatedEvent } from "./erc-721-factory-utils"

// Tests structure (matchstick-as >=0.5.0)
// https://thegraph.com/docs/en/developer/matchstick/#tests-structure-0-5-0

describe("Describe entity assertions", () => {
  beforeAll(() => {
    let name = "Example string value"
    let symbol = "Example string value"
    let maxSupply = BigInt.fromI32(234)
    let contractAddress = Address.fromString(
      "0x0000000000000000000000000000000000000001"
    )
    let sender = Address.fromString(
      "0x0000000000000000000000000000000000000001"
    )
    let newNftContractCreatedEvent = createNftContractCreatedEvent(
      name,
      symbol,
      maxSupply,
      contractAddress,
      sender
    )
    handleNftContractCreated(newNftContractCreatedEvent)
  })

  afterAll(() => {
    clearStore()
  })

  // For more test scenarios, see:
  // https://thegraph.com/docs/en/developer/matchstick/#write-a-unit-test

  test("NftContractCreated created and stored", () => {
    assert.entityCount("NftContractCreated", 1)

    // 0xa16081f360e3847006db660bae1c6d1b2e17ec2a is the default address used in newMockEvent() function
    assert.fieldEquals(
      "NftContractCreated",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "name",
      "Example string value"
    )
    assert.fieldEquals(
      "NftContractCreated",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "symbol",
      "Example string value"
    )
    assert.fieldEquals(
      "NftContractCreated",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "maxSupply",
      "234"
    )
    assert.fieldEquals(
      "NftContractCreated",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "contractAddress",
      "0x0000000000000000000000000000000000000001"
    )
    assert.fieldEquals(
      "NftContractCreated",
      "0xa16081f360e3847006db660bae1c6d1b2e17ec2a-1",
      "sender",
      "0x0000000000000000000000000000000000000001"
    )

    // More assert options:
    // https://thegraph.com/docs/en/developer/matchstick/#asserts
  })
})
